Customer api completed
