+++ pom.xml
@@ -40,6 +40,12 @@
             <artifactId>spring-boot-starter-validation</artifactId>
             <version>2.4.2</version>
         </dependency>
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <version>1.18.18</version>
+            <scope>provided</scope>
+        </dependency>
     </dependencies>
 
     <build>
