+++ src/main/java/com/api/models/Order.java
@@ -1,0 +1,4 @@
+package com.api.models;
+
+public class Order {
+}
