+++ src/main/java/com/api/models/Product.java
@@ -1,5 +1,7 @@
 package com.api.models;
 
+import lombok.Data;
+
 import javax.persistence.*;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
@@ -10,9 +12,10 @@ import java.util.Date;
 
 @Entity
 @Table(name = "products")
+@Data
 public class Product implements Serializable {
     @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
+    @GeneratedValue(strategy = GenerationType.SEQUENCE)
     @Column(name = "product_id")
     private Long id;
 
@@ -29,36 +32,4 @@ public class Product implements Serializable {
 
     @Column(name = "product_date")
     private Date productDate;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
-    }
-
-    public void setPrice(BigDecimal price) {
-        this.price = price;
-    }
-
-    public Date getProductDate() {
-        return productDate;
-    }
-
-    public void setProductDate(Date productDate) {
-        this.productDate = productDate;
-    }
 }
